fix(tasks-redesign): UAT round 7 — uk button overflow, reserves totals tidy, fractional cushion months
UAT #1 — swipe action buttons truncated in uk/pl:
  ACTION_W bumped 88 → 120; reserves-table-row action button width
  matches (w-20 → w-[120px] + px-2). Long localized labels
  ("Виключити" / "Відновити", "Wykluczyć" / "Przywrócić") now fit.

UAT #2 — drop MismatchChip from ReservesTotalsFooter, center on desktop:
  The chip duplicated the new RESERVE_TOPUP / RESERVE_TOPUP_withdraw
  task surface in the per-pill slider. Removed the chip element +
  MismatchChip import. Desktop layout switched
  `sm:justify-between` → `sm:justify-center sm:gap-8` so the Σ rows
  sit centered in the card. `mismatchCents` prop kept on the API for
  source compat — marked `void`.

UAT #3 — "Add category" column wrapping in uk:
  add-category-column width 140 → 160 at base breakpoint so the
  longest localized label ("Додати категорію") stays single-line on
  360–390 px viewports. Desktop unchanged.

UAT #4 + #5 — cushion months: decimal-friendly keyboard + fractional values:

  Migration 0027 (cushion_months_decimal):
    ALTER COLUMN tenancy.budgets.cushion_target_months
      TYPE numeric(4,1) USING ::numeric(4,1).
    Existing integer values round-trip unchanged. CHECK(1..60) still
    applies; fractional values inside the range are now valid.

  Backend math (get-cushion-summary):
    Row type now `string | number | null` (pg returns numeric as
    string). Parse to Number for the response payload; promote to
    BigInt(Math.round(months * 10)) for the BigInt-cents arithmetic
    so requiredCents = (totalCushion × monthsTimes10) / 10n. No
    precision drift on whole-number values; half-month inputs work.

  Zod schema (budget-identity.ts):
    cushion_target_months: z.number().int().min(1).max(60) →
    z.number().min(1).max(60) — drops the integer constraint.

  Frontend (cushion-section.tsx):
    - Input gets `inputMode="decimal"` + `step={0.5}` so iOS / Android
      surfaces the numeric keypad with a decimal key.
    - `parseInt` → `parseFloat` for the computed targetMonths.
    - `Number.isInteger` validation relaxed to `Number.isFinite` so
      4.5 passes the bound check.

  workspace-repo findById row type extended to accept the numeric
  string; mapping coerces to Number with a 6 fallback.

Verification: 38/38 Vitest green on touched suites; 10/10
@tasks-redesign E2E green against live stack; tsc clean. Migration
0027 applied; api + web images rebuilt.

+++ apps/api/src/routes/budget-identity.ts
@@ -28,11 +28,13 @@ const patchBudgetSchema = z.object({
   // lane is exposed in the UI at all. Distinct from cushion_mode_enabled
   // which tracks per-month cushion-vs-normal state via SCD-2 history.
   cushion_enabled: z.boolean().optional(),
-  // Phase 7 Plan 07-07 (D-PH7-15, D-PH7-33): cushion target months multiplier.
-  // Defense in depth: Zod 1..60 here + DB CHECK constraint (1..60) via
-  // migration 0026. Mutating this field triggers recomputeCushionTask after
-  // the identity update.
-  cushion_target_months: z.number().int().min(1).max(60).optional(),
+  // Phase 7 Plan 07-07 (D-PH7-15, D-PH7-33) + UAT round 7: cushion target
+  // months multiplier. Defense in depth: Zod 1..60 here + DB CHECK
+  // constraint (1..60) via migration 0026. Migration 0027 promotes the DB
+  // column to numeric(4,1), so fractional months (e.g. 4.5) are now
+  // permitted — the .int() guard is dropped. Mutating this field triggers
+  // recomputeCushionTask after the identity update.
+  cushion_target_months: z.number().min(1).max(60).optional(),
 });
 
 export function budgetIdentityRoutesFactory(
